Removal of old TODO, @SuppressWarnings which didn't apply anymore, unused imports and variables. Fix some typos in comments and javadoc.
git-svn-id: https://svn.jboss.org/repos/hibernate/search/trunk@15567 1b8cb986-b30d-0410-93ca-fae66ebed9b2

+++ src/java/org/hibernate/search/util/WeakIdentityHashMap.java
@@ -166,7 +166,7 @@ public class WeakIdentityHashMap<K,V> /*extends AbstractMap*/ implements Map<K,V
 	 */
 	public WeakIdentityHashMap() {
 		this.loadFactor = DEFAULT_LOAD_FACTOR;
-		threshold = (int) ( DEFAULT_INITIAL_CAPACITY );
+		threshold = DEFAULT_INITIAL_CAPACITY;
 		table = new Entry[DEFAULT_INITIAL_CAPACITY];
 	}
